GX3-889: Actualizacion de todos los archivos y finalizacion de test
