+++ .github/workflows/CI-Suite.yml
@@ -55,7 +55,7 @@ jobs:
               with:
                   browser: electron
                   command: | #EDITAR AQUÍ EL ARCHIVO SUITE A EJECUTAR:
-                      yarn file cypress/e2e/Tests/Elements/GX3-888-Dinamic.cy.js
+                      yarn file cypress/e2e/Tests/Elements/GX3-889-Dinamic.cy.js
 
             - name: ✅Import Test Results to Xray
               if: always()
