+++ cypress/test-plan/in-sprint/sprint-32/GX3-889.md
@@ -1,0 +1,46 @@
+🎄 ToolsQA | Elements | Dynamic Properties
+
+## Descripción
+
+Como aprendiz de QA,
+
+Quiero probar las Propiedades Dinámicas CUANDO:
+
+-   El elemento tiene un ID dinámico en el DOM
+-   El elemento está habilitado después de 5 segundos
+-   El elemento es visible después de 5 segundos
+-   El color CSS del elemento cambia Para mejorar mis habilidades de prueba en este escenario.
+
+🚩 BUSINESS RULES SPEC
+
+### Elementos:
+
+-   "Este texto tiene un ID aleatorio"
+
+### Para el Elemento: ID Dinámico, el QA debería validar obtener el elemento sin usar "texto" (contiene texto)
+
+El ID es genérico y con cada recarga, el ID cambia, entonces: ¿Cómo obtener un elemento si el ID es dinámico y no tiene clase, nombres u otro
+atributo?
+
+### "Se habilitará en 5 segundos"
+
+El elemento debe ser verificado antes y después de que esté habilitado.
+
+### "Cambio de Color"
+
+El elemento debe ser verificado antes y después de que el color cambie. PD: No hay ninguna propiedad de color CSS asignada al elemento, sino un nombre
+de clase en su lugar.
+
+### "Visible después de 5 segundos"
+
+El elemento debe ser verificado antes y después de que sea visible.
+
+## 🧪 Validaciones
+
+### ✔️ 892|TC1: Validar la capacidad de obtener el elemento "Este texto tiene un ID aleatorio"
+
+### ✔️ 892|TC2: Validar que el elemento esté habilitado después de 5 segundos
+
+### ✔️ 892|TC3: Validar el cambio de estado del elemento "Cambio de Color" después de cinco segundos
+
+### ✔️ 892|TC4: Validar el cambio de estado del elemento "Visible" después de cinco segundos
